Refactor to hook
